+++ packages/cli/src/utils/plugins.ts
@@ -1,5 +1,6 @@
 import fsExtra from 'fs-extra'
-import path from 'path'
+import path from 'node:path'
+import { pathToFileURL } from 'node:url'
 import { withBasePath } from './directory'
 import { logger } from './logger'
 
@@ -69,7 +70,7 @@ export const getPluginsList = async (basePath: string): Promise<Plugin[]> => {
   try {
     const {
       default: { plugins = [] },
-    } = await import(tmpStoreConfigFile)
+    } = await import(pathToFileURL(tmpStoreConfigFile).href)
     return plugins
   } catch (error) {
     logger.error(`Could not load plugins from store config`)
@@ -168,7 +169,9 @@ const generatePluginPages = async (basePath: string, plugins: Plugin[]) => {
     const pluginName = getPluginName(plugin)
     const pluginConfigPath = getPackagePath(pluginName, PLUGIN_CONFIG_FILE)
 
-    const { default: pluginConfig } = await import(pluginConfigPath)
+    const { default: pluginConfig } = await import(
+      pathToFileURL(pluginConfigPath).href
+    )
 
     const { pages: pagesCustom } = getPluginCustomConfig(plugin)
 
@@ -282,7 +285,9 @@ const addPluginsTheme = async (basePath: string, plugins: Plugin[]) => {
         getPackagePath(getPluginName(plugin), 'src', 'themes', 'index.scss')
       )
     )
-    .map((plugin) => `@import "${getPluginName(plugin)}/src/themes/index.scss"`)
+    .map(
+      (plugin) => `@use "${getPluginName(plugin)}/src/themes/index.scss" as *;`
+    )
     .join('\n')
 
   writeFileSync(tmpThemesPluginsFile, pluginImportsContent)
@@ -308,7 +313,9 @@ const generatePluginApis = async (basePath: string, plugins: Plugin[]) => {
     const pluginName = getPluginName(plugin)
     const pluginConfigPath = getPackagePath(pluginName, PLUGIN_CONFIG_FILE)
 
-    const { default: pluginConfig } = await import(pluginConfigPath)
+    const { default: pluginConfig } = await import(
+      pathToFileURL(pluginConfigPath).href
+    )
 
     const { apis: apisCustom } = getPluginCustomConfig(plugin)